Move packages 'doc' and 'print' to internal/pkg.
