+++ internal/pkg/doc/testdata/variables.tf
