+++ internal/pkg/print/print_test.go
@@ -3,8 +3,8 @@ package print_test
 import (
 	"testing"
 
-	"github.com/segmentio/terraform-docs/doc"
-	"github.com/segmentio/terraform-docs/print"
+	"github.com/segmentio/terraform-docs/internal/pkg/doc"
+	"github.com/segmentio/terraform-docs/internal/pkg/print"
 	"github.com/stretchr/testify/assert"
 )
