+++ internal/pkg/print/testdata/json.golden
